+++ openeogeotrellis/utils.py
@@ -20,15 +20,9 @@ import tempfile
 import time
 from functools import partial
 from pathlib import Path
-from typing import Callable, Iterable, Optional, Tuple, Union, Dict, Any, TypeVar, Iterator, TYPE_CHECKING
+from typing import Callable, Iterable, Optional, Tuple, Union, Dict, Any, TypeVar, Iterator
 
-from openeo_driver.integrations.s3.client import S3ClientBuilder as PythonDriverS3ClientBuilder
-from openeo_driver.util.caching import BoundedTtlCache
-
-from openeogeotrellis.integrations.s3proxy.s3_user_context import should_proxy_be_used, build_proxy_s3_client
-
-if TYPE_CHECKING:
-    from mypy_boto3_s3.client import S3Client
+from openeogeotrellis.integrations.s3_client import S3ClientBuilder, eodata_s3_client
 
 import dateutil.parser
 import pyproj
@@ -270,18 +264,6 @@ def set_max_memory(max_total_memory_in_bytes: int):
 
     logger.info("set resource.RLIMIT_AS to {b} bytes".format(b=max_total_memory_in_bytes))
 
-def eodata_s3_client():
-    import boto3
-    aws_access_key_id =  os.environ.get("AWS_ACCESS_KEY_ID")
-    aws_secret_access_key =  os.environ.get("AWS_SECRET_ACCESS_KEY")
-    endpoint = os.environ.get("AWS_S3_ENDPOINT")
-    https = "http" if os.environ.get("AWS_HTTPS").lower() == "no" else "https"
-    s3_client = boto3.client("s3",
-                             aws_access_key_id=aws_access_key_id,
-                             aws_secret_access_key=aws_secret_access_key,
-                             endpoint_url=https + "://" + endpoint)
-    return s3_client
-
 def s3_client():
     # TODO: replace all use cases with get_s3_client(bucket_name)
     # imply a region dependency
@@ -296,38 +278,6 @@ def s3_client():
         aws_secret_access_key=aws_secret_access_key,
         endpoint_url=swift_url)
     return s3_client
-
-
-class S3ClientBuilder:
-    _s3_client_cache = BoundedTtlCache(ttl=12 * 60 * 60, max_size=5)
-
-    @classmethod
-    def from_bucket(cls, bucket_name: str) -> S3Client:
-        """
-        Get an S3 client to allow for interaction with a certain bucket.
-        """
-        return cls._s3_client_cache.get_or_call(bucket_name, lambda: cls._get_s3_client(bucket_name))
-
-    @classmethod
-    def _get_s3_client(cls, bucket_name: str) -> S3Client:
-        if should_proxy_be_used():
-            client = build_proxy_s3_client(bucket_name)
-            if client is None:
-                raise RuntimeError(f"Failed to build proxy S3 client for bucket '{bucket_name}'; see prior warnings.")
-            return client
-        else:
-            return cls._get_direct_s3_client(bucket_name)
-
-    @classmethod
-    def _get_direct_s3_client(cls, bucket_name: str) -> S3Client:
-        """
-        For clients that do not run in a context with OIDC access tokens. eodata is a special case.
-        """
-        if bucket_name.lower() == "eodata":
-            logger.debug("Getting direct S3 client for eodata access")
-            return eodata_s3_client()
-        else:
-            return PythonDriverS3ClientBuilder.from_bucket(bucket_name)
 
 
 def get_s3_file_contents(filename: Union[os.PathLike, str], bucket: Optional[str] = None) -> str: